fix: only the first two guests' fields are filled (#90)
## Problem
Some of the meldeschein form's fields were removed.
This caused the fill script to run into a NullPointer that stopped the
filling of all subsequent fields.

## Solution
The fill script ignores fields that are missing.
They are not removed from the code because they might be added again in
the future.